requirements.txt
pytorch
pandas
numpy
matplotlib